update create store

+++ app/stores/products/index.tsx
@@ -1,28 +1,27 @@
-import React, { useContext, useState, useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import {
   View,
   TouchableOpacity,
   RefreshControl,
-  StatusBar,
   FlatList,
+  Alert,
+  Text,
 } from "react-native";
-import { Ionicons } from "@expo/vector-icons";
+
 import axios from "axios";
 import { config } from "@/constants/config";
-import { AuthContext } from "@/context/auth_context";
 import { useTranslation } from "react-i18next";
 import useFetch from "@/hooks/useFetch";
 import { Toast } from "toastify-react-native";
 import { useRouter } from "expo-router";
-import { SafeAreaView } from "react-native-safe-area-context";
 import Loading from "@/components/ui/Loading";
 import Skeleton from "@/components/ui/Skeleton";
 import NoProductsFound from "@/components/products/NoProductsFound";
 import Header from "@/components/ui/Header";
 import ProductItem from "@/components/products/ProductItem";
 import { useStore } from "@/hooks/useStore";
-import FloadtButton from "@/components/ui/FloadtButton";
 import Layout from "@/components/ui/Layout";
+import FloatButton from "@/components/ui/FloatButton";
 
 interface Product {
   id: number;
@@ -66,23 +65,37 @@ export default function Products() {
   }, [productsData]);
 
   const handleDelete = async (productId: number) => {
-    try {
-      const response = await axios.delete(
-        `${config.URL}/products/${productId}`
+      Alert.alert(
+        t("products.delete_product"),
+        t("products.delete_product_confirmation"),
+        [
+          { text: t("products.cancel"), style: "cancel" },
+          {
+            text: t("products.delete"),
+            style: "destructive",
+            onPress: async () => {
+              try {
+                const response = await axios.delete(
+                  `${config.URL}/products/${productId}`
+                );
+                if (response.data.success) {
+                  Toast.show({
+                    type: "success",
+                    text1: t("products.product_deleted_successfully"),
+                  });
+                  refetchProducts();
+                }
+
+              } catch (error) {
+                Toast.show({
+                  type: "error",
+                  text1: t("products.failed_to_delete_product"),
+                });
+              }
+            },
+          },
+        ]
       );
-      if (response.data.success) {
-        Toast.show({
-          type: "success",
-          text1: t("products.product_deleted_successfully"),
-        });
-        refetchProducts();
-      }
-    } catch (error) {
-      Toast.show({
-        type: "error",
-        text1: t("products.failed_to_delete_product"),
-      });
-    }
   };
 
   const onRefresh = async () => {
@@ -98,6 +111,16 @@ export default function Products() {
     <Layout>
       <Header title={t("products.products")} />
 
+      <View className="flex flex-row justify-between items-center px-5 py-5">
+        <Text className="font-bold"> {t("products.products_count")} - {productsData?.data?.length}</Text>
+
+        <TouchableOpacity 
+          onPress={() => router.push("/stores/products/add")}
+          className="bg-primary rounded-full px-3 py-2">
+          <Text className="text-white">{t("products.add_product")}</Text>
+        </TouchableOpacity>
+      </View>
+
       {loading ? <Loading /> : null}
 
       {productsLoading ? (
@@ -110,7 +133,7 @@ export default function Products() {
 
       {!productsLoading && products.length === 0 ? <NoProductsFound /> : null}
 
-      <View>
+      <View className="mb-10 pb-20">
         <FlatList
           key={"2-columns"}
           data={products}
@@ -130,47 +153,7 @@ export default function Products() {
         />
       </View>
 
-      {/* {loading ? (
-        <Loading />
-      ) : (
-        <>
-          {productsLoading ? (
-            <View className="mt-10 flex gap-4 px-3">
-              {[...Array(5)].map((_, index) => (
-                <Skeleton key={index} height={100} />
-              ))}
-            </View>
-          ) : (
-            <>
-              {products.length === 0 ? (
-                <NoProductsFound />
-              ) : (
-                <>
-                  <FlatList
-                    data={products}
-                    keyExtractor={(item) => item.id.toString()}
-                    refreshControl={
-                      <RefreshControl
-                        refreshing={refreshing}
-                        onRefresh={onRefresh}
-                      />
-                    }
-                    contentContainerStyle={{ padding: 16 }}
-                    renderItem={({ item: product }) => (
-                      <ProductItem
-                        product={product}
-                        categoriesData={categoriesData}
-                        handleDelete={handleDelete}
-                      />
-                    )}
-                  />
-                </>
-              )}
-            </>
-          )}
-          <FloadtButton onPress={() => router.push("/stores/products/add")} />
-        </>
-      )} */}
+       <FloatButton onPress={() => router.push("/stores/products/add")} />
     </Layout>
   );
 }
